move _run_command
