+++ iblvideo/utils.py
@@ -1,33 +1,11 @@
 import logging
 import shutil
-import subprocess
 from pathlib import Path
 
 from one.api import ONE
 from one.remote import aws
 
 _logger = logging.getLogger('ibllib')
-
-
-def _run_command(command: str) -> dict:
-    """
-    Run a shell command using subprocess.
-
-    :param command: command to run
-    :return: dictionary with keys: process, stdout, stderr
-    """
-    process = subprocess.Popen(
-        command,
-        shell=True,
-        stdout=subprocess.PIPE,
-        stderr=subprocess.PIPE,
-    )
-    info, error = process.communicate()
-    return {
-        'process': process,
-        'stdout': info.decode(),
-        'stderr': error.decode(),
-    }
 
 
 def download_and_unzip_file_from_aws(directory, filename, one=None, target_path=None, overwrite=False) -> Path | None:
